Modify: reorganize the directory of utils and libs
