+++ utils/calculateAverage.ts
