+++ utils/getHashIndex.ts
